[MRG] Successive halving (#4)
~~Still very WIP, need feedback on the interface :)~~

I'm now overwriting `_run_search`, taking advantage of the proposed changes in `BaseGridSearch` from https://github.com/scikit-learn/scikit-learn/pull/13145

Note that this forces to remove the `refit` parameter, which has to be set to a custom callable.

`cv_results_` looks good and it's still useful. For e.g. 4 candidates it's going to have 4 (first iter) + 2 (second and last iter) rows.

If you're OK with this new design I'll start to write tests :)